chore: Add handlePress function to KeyboardToolbar for handling tool selection
The handlePress function is added to the KeyboardToolbar component to handle the selection of different tools. If the selected tool is "keyboard" or "star", the setSelectedTool function is called to update the selected tool state. Otherwise, an alert is displayed indicating that the feature is currently under development. This change improves the functionality of the KeyboardToolbar component.